Implement fail-fast logic for dialect resolution, clean-up code around.
